requirements.txt

asgiref==3.8.1
django==5.1.7
sqlparse==0.5.3
mysqlclient==2.2.7
django-simpleui==2025.1.13
Pillow==11.1.0
djangorestframework==3.16.0